pyproject.toml
[project]
name = "pythonweb-10"
version = "0.1.0"
description = ""
authors = [
    {name = "yarqui",email = "y.pelykh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=3.2.0,<4.1.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "cloudinary (>=1.44.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.7"

[tool.pylint.'MESSAGES CONTROL']
disable = "missing-module-docstring, missing-class-docstring, C0116, not-callable, R0913, R0917"

[tool.pylint.main]
source-roots = ["."]